arranjar get n line

+++ so_long.h
@@ -3,19 +3,31 @@
 /*                                                        :::      ::::::::   */
 /*   so_long.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: henrique-reis <henrique-reis@student.42    +#+  +:+       +#+        */
+/*   By: hcarrasq <hcarrasq@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/10 11:55:08 by hcarrasq          #+#    #+#             */
-/*   Updated: 2025/03/24 20:29:37 by henrique-re      ###   ########.fr       */
+/*   Updated: 2025/03/26 15:41:04 by hcarrasq         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef SO_LONG_H
 # define SO_LONG_H
 
+# include <stdio.h>
+# include <unistd.h>
+# include <stdlib.h>
+# include <stdint.h>
+# include <limits.h>
+# include <stdarg.h>
+# include <fcntl.h>
+# include <stdbool.h>
+# include <aio.h>
+# include <stddef.h>
+# include "minilibx-linux/mlx.h"
+# include "libft/libft.h"
+
 #define WIDTH 500
 #define HEIGHT 500
-
 
 
 typedef struct s_cords
@@ -35,34 +47,24 @@ typedef struct s_assets
 
 typedef struct s_map
 {
-	char	**map;
-	int		error;
-	int		width;
-	int		height;
-	int		walls;
-	int		exit;
-	int		collectible;
-	int		player;
+	char		**map;
+	int			error;
+	size_t		width;
+	size_t		height;
+	int			walls;
+	int			exit;
+	int			collectible;
+	int			player;
 }	t_map;
 
 typedef struct s_so_long
 {
-	int	width;
-	int	height;
+	size_t	width;
+	size_t	height;
 	t_map *map;
 	t_assets *ass;
 }	t_so_long;
 
-# include <stdio.h>
-# include <unistd.h>
-# include <stdlib.h>
-# include <stdint.h>
-# include <limits.h>
-# include <stdarg.h>
-# include <fcntl.h>
-# include <stdbool.h>
-# include "minilibx-linux/mlx.h"
-# include "libft/libft.h"
 
 void	map_parsing(char *av, t_map *map, t_assets *assets);
 void	row_checker(int map_fd, t_map *map);
@@ -71,6 +73,6 @@ void	check_assets(char *row, t_map *map, int y, t_assets *assets);
 void	floodfill(t_map *map, t_assets *assets, int x, int y);
 void	ft_error(int panic);
 void	*ft_free(char *str);
-void increment_assets(t_map *map, t_assets *assets, int x, int y);
+void	increment_assets(t_map *map, t_assets *assets, int x, int y);
 
 #endif
